Added unit tests for getClientIp
