--- packages/core/bootstrap/src/lib/util/urlPathBuilder.ts
@@ -1,118 +1,0 @@
-import { getEnv } from '.'
-import { logger } from '../modules'
-
-//  URL Encoding
-const charsToEncode = {
-  ':': '%3A',
-  '/': '%2F',
-  '?': '%3F',
-  '#': '%23',
-  '[': '%5B',
-  ']': '%5D',
-  '@': '%40',
-  '!': '%21',
-  $: '%24',
-  '&': '%26',
-  "'": '%27',
-  '(': '%28',
-  ')': '%29',
-  '*': '%2A',
-  '+': '%2B',
-  ',': '%2C',
-  ';': '%3B',
-  '=': '%3D',
-  '%': '%25',
-  ' ': '%20',
-  '"': '%22',
-  '<': '%3C',
-  '>': '%3E',
-  '{': '%7B',
-  '}': '%7D',
-  '|': '%7C',
-  '^': '%5E',
-  '`': '%60',
-  '\\': '%5C',
-}
-
-/**
- * Check whether the given string contains characters in the given whitelist.
- * @param str The string to check.
- * @param whitelist The string array of whitelist entries. Returns true if any of these are found in 'str', otherwise returns false.
- * @returns {boolean}
- */
-const stringHasWhitelist = (str: string, whitelist: string[]): boolean =>
-  whitelist.some((el) => str.includes(el))
-
-/**
- * Manually iterate through a given string and replace unsafe/reserved characters with encoded values (unless a character is whitelisted)
- * @param str The string to encode.
- * @param whitelist The string array of whitelist entries.
- * @returns {string}
- */
-const percentEncodeString = (str: string, whitelist: string[]): string =>
-  str
-    .split('')
-    .map((char) => {
-      const encodedValue = charsToEncode[char as keyof typeof charsToEncode]
-      return encodedValue && !whitelist.includes(char) ? encodedValue : char
-    })
-    .join('')
-
-/**
- * Build a URL path using the given pathTemplate and params. If a param is found in pathTemplate, it will be inserted there; otherwise, it will be ignored.
- * eg.) pathTemplate = "/from/:from/to/:to" and params = { from: "ETH", to: "BTC", note: "hello" } will become "/from/ETH/to/BTC"
- * @param pathTemplate The path template for the URL path. Each param to include in the path should have a prefix of ':'.
- * @param params The object containing keys & values to be added to the URL path.
- * @param whitelist The list of characters to whitelist for the URL path (if a param contains one of your whitelisted characters, it will not be encoded).
- * @returns {string}
- */
-export const buildUrlPath = (pathTemplate = '', params = {}, whitelist = ''): string => {
-  const allowedChars = whitelist.split('')
-
-  for (const param in params) {
-    const value = params[param as keyof typeof params]
-    if (!value) continue
-
-    // If string contains a whitelisted character: manually replace any non-whitelisted characters with percent encoded values. Otherwise, encode the string as usual.
-    const encodedValue = stringHasWhitelist(value, allowedChars)
-      ? percentEncodeString(value, allowedChars)
-      : encodeURIComponent(value)
-
-    pathTemplate = pathTemplate.replace(`:${param}`, encodedValue)
-  }
-
-  return pathTemplate
-}
-
-/**
- * Build a full URL using the given baseUrl, pathTemplate and params. Uses buildUrlPath to add path & params.
- * @param baseUrl The base URL to add the pathTemplate & params to.
- * @param pathTemplate The path template for the URL path. Leave empty if only searchParams are required.
- * @param params The object containing keys & values to be added to the URL path.
- * @param whitelist The list of characters to whitelist for the URL path.
- * @returns {string}
- */
-export const buildUrl = (baseUrl: string, pathTemplate = '', params = {}, whitelist = ''): string =>
-  new URL(buildUrlPath(pathTemplate, params, whitelist), baseUrl).toString()
-
-//  URL Encoding
-
-let unhandledRejectionHandlerRegistered = false
-
-/**
- * Adapters use to run with Node 14, which by default didn't crash when a rejected promised bubbled up to the top.
- * This function registers a global handler to catch those rejections and just log them to console.
- */
-export const registerUnhandledRejectionHandler = (): void => {
-  if (unhandledRejectionHandlerRegistered) {
-    if (getEnv('NODE_ENV') !== 'test')
-      logger.warn('UnhandledRejectionHandler attempted to be registered more than once')
-    return
-  }
-
-  unhandledRejectionHandlerRegistered = true
-  process.on('unhandledRejection', (reason) => {
-    logger.warn('Unhandled promise rejection reached custom handler')
-    logger.warn(JSON.stringify(reason))
-  })
-}
